+++ MTH2210/Racines_points_fixes/newton_1d.py
@@ -28,7 +28,7 @@ def check_parameters_consistency(f, df, x0, nb_iter, tol_rel, tol_abs, output):
     params_array = [[f,       "f",       types.FunctionType],
                     [df,      "df",      types.FunctionType],
                     [x0,      "x0",      np.float64],
-                    [nb_iter, "nb_iter", np.int],
+                    [nb_iter, "nb_iter", int],
                     [tol_rel, "tol_rel", np.float64],
                     [tol_abs, "tol_abs", np.float64],
                     [output,  "output",  str]]
@@ -132,12 +132,12 @@ def newton_1d(f, df, x0, nb_iter=100, tol_rel=10**-8, tol_abs=10**-8, output="")
     """Méthode de recherche d'une racine de la fonction scalaire f via la méthode de Newton :
         - x_0 donné,
         - x_kp1 = xk - f(xk)/f'(xk).
-    
+
     Les arguments attendus sont :
         - une fonction  f, admettant en entrée un scalaire x et renvoyant un scalaire f(x),
         - une fonction df, admettant en entrée un scalaire x et renvoyant un scalaire f'(x),
         - un scalaire  x0 (de type int, float ou np.float64), point de départ de la méthode itérative.
-    
+
     Les arguments optionnels sont :
         - un entier nb_iter (défaut = 100 ) définissant le nombre maximal d'itérations allouées à la méthode,
         - un réel   tol_rel (défaut = 1e-8) définissant la condition d'arrêt abs(x_k-x_km1) / (abs(x_k)+eps) <= tol_rel,
@@ -151,32 +151,30 @@ def newton_1d(f, df, x0, nb_iter=100, tol_rel=10**-8, tol_abs=10**-8, output="")
          - f est définie en x0, et renvoie un scalaire,
         - df est définie en x0, et renvoie un scalaire,
         - tous les paramètres reçus ont bien le type attendu.
-    
+
     Les sorties de la méthode sont :
         - list_x, la liste des points x_k,
         - list_f, les valeurs par  f des éléments de list_x,
         - list_d, les valeurs par df des éléments de list_x.
-        
+
     Exemples d'appel :
         - newton_1d(lambda x : np.sin(x), lambda x:np.cos(x), 1),
         - newton_1d(lambda x :x**2, lambda x:2*x, 1).
     """
-    
+
     # Test des paramètres et définition de la destination de sortie des itérations
     check_parameters_consistency(f, df, x0, nb_iter, tol_rel, tol_abs, output)
     write_iter, write_stopping = writing_function.define_writing_function(format_iter, output)
-    
+
     # Initialisation de l'algorithme
     k, list_x, list_f, list_d = init_algo(f, df, x0)
     write_iter(k, list_x, list_f, list_d)
-    
+
     # Déroulement de l'algorithme
     while not(stopping_criteria(k, list_x, list_f, list_d, nb_iter, tol_rel, tol_abs)[0]):
         k, list_x, list_f, list_d = iter_algo(f, df, k, list_x, list_f, list_d)
         write_iter(k, list_x, list_f, list_d)
-    
+
     write_stopping(stopping_criteria(k, list_x, list_f, list_d, nb_iter, tol_rel, tol_abs)[1])
     # Renvoi de la liste des approximations de la racine, des valeurs de f associées, et des erreurs relatives
     return(list_x, list_f, list_d)
-
-
