+++ MTH2210/Racines_points_fixes/newton_nd_avec_der.py
@@ -28,7 +28,7 @@ def check_parameters_consistency(f, jac, x0, nb_iter, tol_rel, tol_abs, output):
     params_array = [[f,       "f",       types.FunctionType],
                     [jac,     "jac",     types.FunctionType],
                     [x0,      "x0",      np.ndarray],
-                    [nb_iter, "nb_iter", np.int],
+                    [nb_iter, "nb_iter", int],
                     [tol_rel, "tol_rel", np.float64],
                     [tol_abs, "tol_abs", np.float64],
                     [output,  "output",  str]]
@@ -141,12 +141,12 @@ def newton_nd_avec_der(f, jac, x0, nb_iter=100, tol_rel=10**-8, tol_abs=10**-8, 
     """Méthode de recherche d'une racine de la fonction vectorielle f via la méthode de Newton avec approximation de la jacobienne :
         - x_0 donné,
         - x_kp1 = xk - Jac(f)(x_k)^-1*f(x_k).
-    
+
     Les arguments attendus sont :
         - une fonction   f, admettant en entrée un vecteur x et renvoyant un  vecteur f(x),
         - une fonction jac, admettant en entrée un vecteur x et renvoyant une matrice Jac(f)(x),
         - un scalaire   x0 (de type int, float ou np.float64), point de départ de la méthode itérative.
-    
+
     Les arguments optionnels sont :
         - un entier nb_iter (défaut = 100 ) définissant le nombre maximal d'itérations allouées à la méthode,
         - un réel   tol_rel (défaut = 1e-8) définissant la condition d'arrêt abs(x_k-x_km1) / (abs(x_k)+eps) <= tol_rel,
@@ -160,12 +160,12 @@ def newton_nd_avec_der(f, jac, x0, nb_iter=100, tol_rel=10**-8, tol_abs=10**-8, 
          -   f est définie en x0, et renvoie un  vecteur de même dimension que x0,
          - jac est définie en x0, et renvoie une matrice carrée de dimension dim(x0)*dim(x0),
          - tous les paramètres reçus ont bien le type attendu.
-    
+
     Les sorties de la méthode sont :
         - list_x, la liste des points x_k,
         - list_f, les valeurs par     f  des éléments de list_x,
         - list_d, les valeurs par Jac(f) des éléments de list_x.
-        
+
     Exemples d'appel :
         - newton_nd(lambda x:x**2, lambda x: np.array([2*x]), np.array([1,1,1])),
         - def f(x):
@@ -175,22 +175,20 @@ def newton_nd_avec_der(f, jac, x0, nb_iter=100, tol_rel=10**-8, tol_abs=10**-8, 
           x0 = np.array([1,1,1])
           newton_nd_avec_der(f, jac, x0).
     """
-    
+
     # Test des paramètres et définition de la destination de sortie des itérations
     check_parameters_consistency(f, jac, x0, nb_iter, tol_rel, tol_abs, output)
     write_iter, write_stopping = writing_function.define_writing_function(format_iter, output)
-    
+
     # Initialisation de l'algorithme
     k, list_x, list_f, list_d = init_algo(f, jac, x0)
     write_iter(k, list_x, list_f)
-    
+
     # Déroulement de l'algorithme
     while not(stopping_criteria(k, list_x, list_f, list_d, nb_iter, tol_rel, tol_abs)[0]):
         k, list_x, list_f, list_d = iter_algo(f, jac, k, list_x, list_f, list_d)
         write_iter(k, list_x, list_f)
-    
+
     write_stopping(stopping_criteria(k, list_x, list_f, list_d, nb_iter, tol_rel, tol_abs)[1])
     # Renvoi de la liste des approximations de la racine, des valeurs de f associées, et des erreurs relatives
     return(list_x, list_f, list_d)
-
-
